feat: 出力構造の統一と実RGB SLCSデータ生成・評価基盤

+++ src/tasks/blcs/data/types.py
@@ -140,8 +140,8 @@ class BLCSSceneMeta:
 
     shots: list[dict]  # List of BLCSShotEventMeta.to_dict() results
 
-    fps_out: int
-    sim_fps: int
+    fps_out: float
+    sim_fps: float
     num_frames: int
     num_cameras_sampled: int
     num_cameras: int
